# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Notebook API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Notebook/v1 API"
        }
    ],
    "paths": {
        "/notebook": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Покозать все записи",
                "operationId": "AllRecords",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Номер страницы",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все хорошо"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Создать запись",
                "operationId": "RecordCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone",
                                    "email"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "ФИО"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "79999999999"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "email@example.com"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Компания"
                                    },
                                    "photo": {
                                        "description": "Upload image",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Созданно"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/notebook/{id}": {
            "get": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Получить запись по ID",
                "operationId": "RecordGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все хорошо"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Изменить запись по ID",
                "operationId": "RecordUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": ""
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "full_name",
                                    "phone"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "string",
                                        "example": "ФИО(update)"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "79999999999(update)"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "email@example.com(update)"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Компания(update)"
                                    },
                                    "date_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "01-01-1111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Все хорошо"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notebook"
                ],
                "summary": "Удалить запись по ID",
                "operationId": "RecordDelete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Удалить по ID ",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все хорошо"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Notebook": {
                "title": "Notebook",
                "description": "Notebook model",
                "xml": {
                    "name": "Notebook"
                }
            }
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Notebook",
            "description": "Notebook pages"
        }
    ]
}